requirements.txt
djangorestframework
markdown
django-filter
httpie
drf-yasg